+++ src/main/java/absolutelyaya/ultracraft/registry/ItemRegistry.java
@@ -230,6 +230,8 @@ public class ItemRegistry
 			content.add(StainedGlassWindowItem.getStack(false));
 			content.add(StainedGlassWindowItem.getStack(true));
 			content.add(BlockRegistry.SLAB_BLOCK.asItem());
+			content.add(BlockRegistry.CARPET.asItem());
+			content.add(BlockRegistry.FLOWERBED.asItem());
 			content.add(BLOOD_BUCKET);
 			content.add(PIERCE_REVOLVER);
 			content.add(MARKSMAN_REVOLVER);